pyproject.toml
[tool.poetry]
name = "kithon"
readme = "README.md"
version = "0.0.1"
description = "transpiler python into other language"
authors = ["Aleksey Ploskov"]
license = "MIT"
repository = "https://github.com/alploskov/kithon"
packages = [{ include = "kithon"}]
include = [
	"translators/js/*",
	"translators/js/libs/*",
	"translators/js/objects/*",

	"translators/go/*",
	"translators/go/libs/*",
	"translators/go/objects/*"
]

[tool.poetry.dependencies]
python = ">=3.9.0"
Jinja2 = ">=2.11.2,<4.0.0"
PyYaml = ">=5.4.0"
typer = "^0.4.0"

[tool.poetry.scripts]
kithon = "kithon.cli:tpy"
kithon-gen = "kithon.cli:tpy_gen"
kithon-new = "kithon.cli:tpy_new"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"